Made SimpleServer class
Made a class for server obj focusing on trying to use function callbacks rather than lambdas so it can be a little more cleaner code (especially for the part where we are matching url routes and request methods). Next step is trying to implement a datastructure (maybe dictionary) of objects that allow for easy matching of urls. The objs in the dictionary will belong to a different route and have their own callbacks for each of the basic http methods